Rewrite w/ classes
